+++ wire-ios/Wire-iOS/Sources/Developer/DeveloperTools/DebugActions/DeveloperDebugActionsView.swift
@@ -19,38 +19,35 @@
 import SwiftUI
 
 struct DeveloperDebugActionsView: View {
-    
+
     @ObservedObject var viewModel: DeveloperDebugActionsViewModel
     @State var userInput: String = ""
-    
+
     var body: some View {
         List(viewModel.buttons) { button in
             Button(action: button.action) {
                 Text(button.title)
             }
         }
-        .sheet(item: $viewModel.mlsGroupSearchItem,content: mlsGroupSearchView)
+        .sheet(item: $viewModel.mlsGroupSearchItem, content: mlsGroupSearchView)
         .presentationDetents([.medium])
         .presentationDragIndicator(.visible)
     }
-    
+
     @ViewBuilder
     func mlsGroupSearchView(_ item: MLSGroupSearchItem) -> some View {
         List {
             Section("Conversations with MLS Group ID") {
                 HStack {
                     TextField("Enter MLS Group ID", text: $userInput)
-                    Button("Search") {
-                        Task {
-                            await viewModel.findConversations(with: userInput.trim())
-                        }
-                    }
-                    .disabled(userInput.isEmpty)
+                        .onSubmit(submitUserInput)
+                    Button("Search", action: submitUserInput)
+                        .disabled(userInput.isEmpty)
                 }
             }
             switch item {
-            case .result(let results):
-                if results.isEmpty && !userInput.isEmpty {
+            case let .result(results):
+                if results.isEmpty || userInput.isEmpty {
                     Section {
                         Text("Nothing found")
                     }
@@ -72,28 +69,10 @@ struct DeveloperDebugActionsView: View {
             }
         }
     }
-}
 
-
-struct ResultView: View {
-    var result: ConversationResult
-    
-    var body: some View {
-        VStack(alignment: .leading) {
-            Text("Name:")
-                .bold()
-            Text(result.name)
-            
-            Text("Conversation id:")
-                .bold()
-            Text(result.id)
-            
-            Text("MLS Group id:")
-                .bold()
-            Text(result.groupID?.description ?? "-")
-        }
-        .onTapGesture {
-            UIPasteboard.general.string = result.id
+    private func submitUserInput() {
+        Task {
+            await viewModel.findConversations(with: userInput.trim())
         }
     }
 }
